(web admin) abstract into separate pfNetworkGraph component
